requirements.txt
-e git+https://github.com/entone/envy.git#egg=envy
-e git+https://github.com/entone/Humongolus.git#egg=Humongolus
pyzmq
simplejson
pycrypto==2.6
unidecode
boto
pyserial